requirements.txt
certifi==2023.11.17
charset-normalizer==3.3.2
GPUtil==1.4.0
idna==3.6
psutil==5.9.7
py-cpuinfo==9.0.0
requests==2.31.0
screeninfo==0.8.1
speedtest-cli==2.1.3
urllib3==2.1.0
